All Codes/pubsub_to_firestore/requirements.txt
google-cloud-firestore
cloudevents
